Revert "exploring breaking down signed module into multiple files"
This reverts commit 90051b6500b458a4ef55c7eb30a1a7fe56ec8873.
